Layout folder moved to src
