+++ src/Layout/PublicLayout.js
@@ -1,8 +1,8 @@
 import React from "react";
-import Header from "../Header/Header";
-import Footer from "../Footer/Footer";
+import Header from "../Components/Header/Header";
+import Footer from "../Components/Footer/Footer";
 import { BrowserRouter, Route, Switch } from "react-router-dom";
-import RegisterForm from "../Auth/RegisterForm";
+import RegisterForm from "../Components/Auth/RegisterForm";
 
 const PublicLayout = ({ children }) => {
   return (
